fix: use `raw_sql_to_block_number` and its complement, fix the last test case due to a type representation issue

+++ crates/store/src/db/schema.rs
@@ -4,7 +4,7 @@ diesel::table! {
     account_deltas (account_id, block_num) {
         account_id -> Binary,
         block_num -> BigInt,
-        nonce -> Integer, // TODO it's a `Felt`, so why is this not a `BigInt`?
+        nonce -> BigInt,
     }
 }
